add token authentication requirement

+++ part-4/router/apifunctions/apifunctions.go
@@ -2,9 +2,11 @@ package apifunctions
 
 import (
 	"encoding/json"
+	"errors"
 	"io/ioutil"
 	"net/http"
 
+	"github.com/dgrijalva/jwt-go"
 	"github.com/gorilla/mux"
 	"github.corp.globant.com/diego-maranges/GolangBootcamp/part-4/db"
 )
@@ -35,12 +37,56 @@ func response(w http.ResponseWriter, status int, results interface{}) {
 	json.NewEncoder(w).Encode(results)
 }
 
+func checkIfLogIn(w http.ResponseWriter, r *http.Request) int {
+	// We can obtain the session token from the requests cookies, which come with every request
+	c, err := r.Cookie("token")
+	if err != nil {
+		if err == http.ErrNoCookie {
+			// If the cookie is not set, return an unauthorized status
+			return http.StatusUnauthorized
+		}
+		// For any other type of error, return a bad request status
+		return http.StatusBadRequest
+	}
+
+	// Get the JWT string from the cookie
+	tknStr := c.Value
+
+	// Initialize a new instance of `Claims`
+	claims := &Claims{}
+
+	// Parse the JWT string and store the result in `claims`.
+	// Note that we are passing the key in this method as well. This method will return an error
+	// if the token is invalid (if it has expired according to the expiry time we set on sign in),
+	// or if the signature does not match
+	tkn, err := jwt.ParseWithClaims(tknStr, claims, func(token *jwt.Token) (interface{}, error) {
+		return jwtKey, nil
+	})
+	if err != nil {
+		if err == jwt.ErrSignatureInvalid {
+			return http.StatusUnauthorized
+		}
+		return http.StatusBadRequest
+	}
+	if !tkn.Valid {
+		return http.StatusUnauthorized
+	}
+
+	return http.StatusOK
+}
+
 /*CreateNewCar return
 
 status 201 if the car is already exist*/
 func CreateNewCar(w http.ResponseWriter, r *http.Request) {
-	if _, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], true); err != nil {
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	if _, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], true); err != nil {
 		errorResponse(w, http.StatusCreated, err)
+		return
 	}
 	response(w, http.StatusOK, nil)
 }
@@ -51,13 +97,20 @@ status 404 if the car does not exist,
 
 status 409 if have any error traing to load the file or read the Map*/
 func ReturnCar(w http.ResponseWriter, r *http.Request) {
-	dataBase, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	dataBase, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := dataBase.LoadFile(); er != nil {
 		errorResponse(w, http.StatusConflict, er)
+		return
 	}
 
 	carMap, err := dataBase.ReturnMap()
@@ -72,13 +125,20 @@ func ReturnCar(w http.ResponseWriter, r *http.Request) {
 
 status 404 if the car does not exist or the car does not deleted*/
 func DeleteCar(w http.ResponseWriter, r *http.Request) {
-	myDB, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	myDB, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := myDB.DeleteFile(); er != nil {
 		errorResponse(w, http.StatusNotFound, er)
+		return
 	}
 
 	response(w, http.StatusOK, nil)
@@ -90,13 +150,20 @@ status 404 if the car does not exist or the car does not have this item
 
 status 409 if have any error load data*/
 func ReturnItem(w http.ResponseWriter, r *http.Request) {
-	dataBase, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	dataBase, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := dataBase.LoadFile(); er != nil {
 		errorResponse(w, http.StatusConflict, er)
+		return
 	}
 
 	itemID := mux.Vars(r)[itemPath]
@@ -114,17 +181,24 @@ status 404 if the car does not exist
 
 status 409 if have any error load/save data, adding the item or create Json response*/
 func AddItem(w http.ResponseWriter, r *http.Request) {
-	dataBase, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	dataBase, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := dataBase.LoadFile(); er != nil {
 		errorResponse(w, http.StatusConflict, er)
+		return
 	}
 
 	itemID := mux.Vars(r)[itemPath]
-	if er := dataBase.Add(directoyDB, itemID); er != nil {
+	if er := dataBase.Add(itemID); er != nil {
 		errorResponse(w, http.StatusConflict, er)
 		return
 	}
@@ -150,13 +224,20 @@ status 409 if have any error load/save data, updating the item or create/read Js
 
 status 400 if Json received is wrong*/
 func UpdateItem(w http.ResponseWriter, r *http.Request) {
-	dataBase, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	dataBase, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := dataBase.LoadFile(); er != nil {
 		errorResponse(w, http.StatusConflict, er)
+		return
 	}
 
 	itemID := mux.Vars(r)[itemPath]
@@ -196,13 +277,20 @@ status 404 if the car does not exist
 
 status 409 if have any error load/save data, erasing the item or create Json response*/
 func DeleteItem(w http.ResponseWriter, r *http.Request) {
-	dataBase, err := db.CreateNewDBInstance(mux.Vars(r)[carPath], false)
+	if status := checkIfLogIn(w, r); status != 200 {
+		errorResponse(w, status, errors.New("error login"))
+		return
+	}
+
+	dataBase, err := db.CreateNewDBInstance(directoyDB, mux.Vars(r)[carPath], false)
 	if err != nil {
 		errorResponse(w, http.StatusNotFound, err)
+		return
 	}
 
 	if er := dataBase.LoadFile(); er != nil {
 		errorResponse(w, http.StatusConflict, er)
+		return
 	}
 
 	itemID := mux.Vars(r)[itemPath]
